+++ src/components/form/index.ts
@@ -1,0 +1,2 @@
+export { Input } from './input';
+export { ValidationError } from './validation-error';
